identity/requirements.txt
asgiref==3.6.0
autopep8==2.0.1
Django==4.1.5
django-rest-framework==0.1.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
dnspython==2.3.0
mysqlclient==2.1.1
pipdeptree==2.3.3
pycodestyle==2.10.0
PyJWT==2.6.0
pytz==2022.7.1
sqlparse==0.4.3
